[fontc] Move Config to new 'config' module
This also makes the 'init' function into a method on this type, but
otherwise avoids any substantive changes.
